+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Latest
 - Added weekly news, showing all the statistics for a given week.
+- Added monthly news, showing all the statistics for a given month.
 
 ## 2025.05.16
 - Changed the News page so it shows news for the past 7 days, not the current week.